clean: rename controllers and routes
